Try to rework OTel initialization for Lambda
Honestly, this is a mess thanks to how hard I'm trying to fall back
gracefully, and thanks to Go's shortcomings as a language.

I think I should construct the "resource" separately from the provider.
Most of the issue seems to be interleaving those two things, and having
multiple return paths that construct a TracerProvider.